Refactor: Saparate header for `RecordSectionsTracker`
