+++ compiler-rt/lib/orc/macho_platform.cpp
@@ -17,6 +17,7 @@
 #include "error.h"
 #include "interval_map.h"
 #include "jit_dispatch.h"
+#include "sections_tracker.h"
 #include "wrapper_function_utils.h"
 
 #include <algorithm>
